Tweaks for warnings and building under RedHat 5 and Ubuntu 13 components/canny.rcc/canny.c: warning components/dds_complex.rcc/dds_complex.c: warning components/file_write.rcc/file_write.c: warning core/container/interfaces/src/OcpiContainerPort.cxx: warning core/dataplane/msg_drivers/dds/src/OcpiDDSTransport.cxx: warning jkamd_env_linux_x86_64.sh: remove opensplice for RHEL5 tests/facilities/test_os/tests/src/TestOcpiOsProcessorManager.cxx: warning core/library/src/OcpiLibraryManager.cxx: missing strict header for ubuntu tools/cdk/examples/cat_xml/catHello.cxx: ditto tools/cdk/examples/ext2file_xml/ext2file.cxx: ditto tools/cdk/examples/fsk_modem_xml/fsk_modem.cxx: ditto tools/cdk/examples/hello/hello.cxx: ditto tools/cdk/examples/hello_xml/appHello.cxx:ditto tools/cdk/export/bin/linux-u13-x86_64/ocpigen: fix bad link for ubuntu tools/cdk/include/ocpisetup.mk: remove unnecessary (and broken on ubuntu) -luuid, add -pthread for ubuntu

+++ components/file_write.rcc/file_write.c
@@ -17,7 +17,7 @@ typedef struct {
   int fd;
   int started;
 } MyState;
-static uint32_t mysizes[] = {sizeof(MyState), 0};
+static size_t mysizes[] = {sizeof(MyState), 0};
 
 FILE_WRITE_METHOD_DECLARATIONS;
 RCCDispatch file_write = {
